+++ crates/cli-sub-agent/src/run_cmd_post.rs
@@ -2,19 +2,32 @@
 //! update, and seed-session management.
 //!
 //! Extracted from `run_cmd.rs` to keep module sizes manageable.
+//!
+//! Failover evaluation (rate-limit / transport-error → tier failover) lives
+//! in [`run_cmd_post_failover`](../run_cmd_post_failover/index.html) and is
+//! re-exported below to keep existing call sites stable.
 
-use std::{path::Path, time::Duration};
+use std::path::Path;
 
 use anyhow::Result;
 use tracing::{debug, info, warn};
 
 use csa_config::ProjectConfig;
-use csa_core::types::{FallbackAttempt, ToolName};
+use csa_core::types::ToolName;
 use csa_scheduler::FallbackChain;
 use csa_session::{PhaseEvent, SessionPhase, load_result, save_result};
 
 use crate::run_cmd_fork::{ForkResolution, load_child_return_packet};
 use crate::run_cmd_tool_selection::resolve_slot_wait_timeout_seconds;
+
+#[path = "run_cmd_post_failover.rs"]
+mod failover;
+pub(crate) use failover::{
+    RateLimitAction, detect_permanent_tool_exhaustion_result,
+    detect_permanent_tool_exhaustion_text, evaluate_error_rate_limit_failover,
+    evaluate_rate_limit_failover, format_tool_exhausted_summary,
+    is_permanent_tool_exhaustion_error,
+};
 
 /// Handle the fork-call parent resume protocol after child execution completes.
 ///
@@ -141,34 +154,6 @@ pub(crate) fn update_fork_genealogy(
     }
 }
 
-/// Outcome of rate-limit failover evaluation.
-pub(crate) enum RateLimitAction {
-    /// No rate limit detected; break with result.
-    NoRateLimit,
-    /// Rate limit detected but no failover possible; break with result.
-    ExhaustedFailovers,
-    /// Retry with a different tool.
-    Retry {
-        new_tool: ToolName,
-        new_model_spec: Option<String>,
-    },
-}
-
-#[derive(Clone, Copy)]
-enum TransportErrorFailoverKind {
-    RateLimit,
-    AcpCrashRetryExhausted,
-    GeminiRetryChainExhausted,
-}
-
-struct TransportErrorFailoverSignal {
-    kind: TransportErrorFailoverKind,
-    matched_pattern: String,
-    reason: String,
-    quota_exhausted: bool,
-    requires_init_failure_window: bool,
-}
-
 /// Overwrite a session's `result.toml` after a post-exec gate failure (#1486).
 ///
 /// The result was written with status=success before the gate ran.
@@ -205,498 +190,6 @@ pub(crate) fn overwrite_result_as_post_exec_gate_failure(
                 error = %load_err,
                 "Failed to load result.toml for post-exec gate failure overwrite"
             );
-        }
-    }
-}
-
-pub(crate) fn format_tool_exhausted_summary(tool_name: &str, matched_pattern: &str) -> String {
-    format!(
-        "tool_exhausted: {tool_name} permanent quota exhaustion detected \
-         (matched '{matched_pattern}'); no retry or tool fallback attempted. \
-         Inspect the tool account billing/quota or choose another tool explicitly."
-    )
-}
-
-pub(crate) fn detect_permanent_tool_exhaustion_result(
-    tool_name_str: &str,
-    exec_result: &csa_process::ExecutionResult,
-    current_model_spec: Option<&str>,
-) -> Option<csa_scheduler::RateLimitDetected> {
-    detect_permanent_tool_exhaustion_text(
-        tool_name_str,
-        &exec_result.summary,
-        &exec_result.stderr_output,
-        &exec_result.output,
-        exec_result.exit_code,
-        current_model_spec,
-    )
-}
-
-pub(crate) fn detect_permanent_tool_exhaustion_text(
-    tool_name_str: &str,
-    summary: &str,
-    stderr: &str,
-    stdout: &str,
-    exit_code: i32,
-    current_model_spec: Option<&str>,
-) -> Option<csa_scheduler::RateLimitDetected> {
-    if exit_code == 0 {
-        return None;
-    }
-    let stdout_with_summary = if summary.trim().is_empty() {
-        stdout.to_string()
-    } else if stdout.trim().is_empty() {
-        summary.to_string()
-    } else {
-        format!("{summary}\n{stdout}")
-    };
-    csa_scheduler::detect_rate_limit(
-        tool_name_str,
-        stderr,
-        &stdout_with_summary,
-        exit_code,
-        current_model_spec,
-    )
-    .filter(|detected| detected.quota_exhausted)
-}
-
-pub(crate) fn is_permanent_tool_exhaustion_error(
-    tool_name_str: &str,
-    error_message: &str,
-    current_model_spec: Option<&str>,
-) -> bool {
-    detect_transport_error_failover_signal(tool_name_str, error_message, current_model_spec)
-        .is_some_and(|signal| signal.quota_exhausted)
-}
-
-fn detect_transport_error_failover_signal(
-    tool_name_str: &str,
-    error_message: &str,
-    current_model_spec: Option<&str>,
-) -> Option<TransportErrorFailoverSignal> {
-    let error_lower = error_message.to_ascii_lowercase();
-
-    if error_lower.contains("acp crash retry exhausted")
-        || error_lower.contains("crash retry exhausted")
-    {
-        let matched_pattern = if error_lower.contains("acp crash retry exhausted") {
-            "acp crash retry exhausted"
-        } else {
-            "crash retry exhausted"
-        };
-        return Some(TransportErrorFailoverSignal {
-            kind: TransportErrorFailoverKind::AcpCrashRetryExhausted,
-            matched_pattern: matched_pattern.to_string(),
-            reason: "acp_crash_retry_exhausted".to_string(),
-            quota_exhausted: false,
-            requires_init_failure_window: false,
-        });
-    }
-
-    if error_lower.contains("gemini acp retry chain exhausted")
-        || error_lower.contains("retry chain exhausted")
-    {
-        let matched_pattern = if error_lower.contains("gemini acp retry chain exhausted") {
-            "gemini acp retry chain exhausted"
-        } else {
-            "retry chain exhausted"
-        };
-        return Some(TransportErrorFailoverSignal {
-            kind: TransportErrorFailoverKind::GeminiRetryChainExhausted,
-            matched_pattern: matched_pattern.to_string(),
-            reason: "gemini_retry_chain_exhausted".to_string(),
-            quota_exhausted: csa_core::gemini::detect_permanent_quota_exhaustion_pattern(
-                error_message,
-            )
-            .is_some(),
-            requires_init_failure_window: false,
-        });
-    }
-
-    csa_scheduler::detect_rate_limit(
-        tool_name_str,
-        error_message,
-        "",
-        1, // synthetic non-zero exit code
-        current_model_spec,
-    )
-    .map(|rate_limit| {
-        let requires_init_failure_window = csa_scheduler::requires_init_failure_window(&rate_limit);
-        TransportErrorFailoverSignal {
-            kind: TransportErrorFailoverKind::RateLimit,
-            matched_pattern: rate_limit.matched_pattern,
-            reason: rate_limit.reason,
-            quota_exhausted: rate_limit.quota_exhausted,
-            requires_init_failure_window,
-        }
-    })
-}
-
-fn allows_init_failure_failover(
-    tool_name: &str,
-    reason: &str,
-    requires_init_failure_window: bool,
-    attempt_elapsed: Option<Duration>,
-) -> bool {
-    if !requires_init_failure_window {
-        return true;
-    }
-    let Some(elapsed) = attempt_elapsed else {
-        return true;
-    };
-    if csa_scheduler::within_init_failure_window(elapsed) {
-        warn!(
-            tool = %tool_name,
-            reason = %reason,
-            elapsed_ms = elapsed.as_millis(),
-            "[csa-failover] {tool_name} failed with {reason}, falling back to next tier model"
-        );
-        return true;
-    }
-    warn!(
-        tool = %tool_name,
-        reason = %reason,
-        elapsed_ms = elapsed.as_millis(),
-        "[csa-failover] HTTP failure occurred after initialization window; not attempting automatic tier fallback"
-    );
-    false
-}
-
-/// Check for 429 rate-limit signals and decide whether to failover.
-///
-/// Returns `RateLimitAction` to drive `continue`/`break` in the caller loop.
-/// Appends a `FallbackAttempt` to `fallback_chain` when a retry is triggered.
-///
-/// `resolved_tier_name` and `tried_specs` enable intra-tier failover: when the
-/// caller is running under a named tier, we pass spec-level exclusion so that
-/// a different model within the same tier can be selected.
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn evaluate_rate_limit_failover(
-    tool_name_str: &str,
-    exec_result: &csa_process::ExecutionResult,
-    attempts: usize,
-    max_failover_attempts: usize,
-    tried_tools: &mut Vec<String>,
-    tried_specs: &mut Vec<String>,
-    tier_auto_select: bool,
-    resolved_tier_name: Option<&str>,
-    executed_session_id: Option<&str>,
-    effective_session_arg: Option<&str>,
-    ephemeral: bool,
-    prompt_text: &str,
-    project_root: &Path,
-    config: Option<&ProjectConfig>,
-    task_needs_edit: Option<bool>,
-    current_model_spec: Option<&str>,
-    fallback_chain: &mut FallbackChain,
-    attempt_elapsed: Option<Duration>,
-) -> Result<RateLimitAction> {
-    let rate_limit = match csa_scheduler::detect_rate_limit(
-        tool_name_str,
-        &exec_result.stderr_output,
-        &format!("{}\n{}", exec_result.summary, exec_result.output),
-        exec_result.exit_code,
-        current_model_spec,
-    ) {
-        Some(rl) => rl,
-        None => return Ok(RateLimitAction::NoRateLimit),
-    };
-
-    if rate_limit.quota_exhausted {
-        warn!(
-            tool = %tool_name_str,
-            pattern = %rate_limit.matched_pattern,
-            "Permanent tool quota exhaustion detected; not attempting runtime failover"
-        );
-        return Ok(RateLimitAction::ExhaustedFailovers);
-    }
-
-    if !allows_init_failure_failover(
-        tool_name_str,
-        &rate_limit.reason,
-        csa_scheduler::requires_init_failure_window(&rate_limit),
-        attempt_elapsed,
-    ) {
-        return Ok(RateLimitAction::NoRateLimit);
-    }
-
-    if !tier_auto_select {
-        return Ok(RateLimitAction::NoRateLimit);
-    }
-
-    info!(
-        tool = %tool_name_str,
-        pattern = %rate_limit.matched_pattern,
-        quota_exhausted = rate_limit.quota_exhausted,
-        attempt = attempts,
-        max = max_failover_attempts,
-        "Rate limit detected, attempting failover"
-    );
-
-    if attempts >= max_failover_attempts {
-        warn!(
-            "Max failover attempts ({}) reached, returning error",
-            max_failover_attempts
-        );
-        return Ok(RateLimitAction::ExhaustedFailovers);
-    }
-
-    tried_tools.push(tool_name_str.to_string());
-    if let Some(spec) = current_model_spec {
-        tried_specs.push(spec.to_string());
-    }
-
-    // Prefer the actually-executed session (important for forks where
-    // effective_session_arg starts as None) so decide_failover evaluates
-    // the fork session's context, not the parent session.
-    let failover_session_ref = executed_session_id.or(effective_session_arg);
-    let session_state = if !ephemeral {
-        failover_session_ref.and_then(|sid| {
-            let sessions_dir = csa_session::get_session_root(project_root)
-                .ok()?
-                .join("sessions");
-            let resolved_id = csa_session::resolve_session_prefix(&sessions_dir, sid).ok()?;
-            csa_session::load_session(project_root, &resolved_id).ok()
-        })
-    } else {
-        None
-    };
-
-    let task_needs_edit = task_needs_edit
-        .or_else(|| crate::run_helpers::infer_task_edit_requirement(prompt_text))
-        .or_else(|| config.map(|cfg| cfg.can_tool_edit_existing(tool_name_str)));
-
-    let Some(cfg) = config else {
-        return Ok(RateLimitAction::ExhaustedFailovers);
-    };
-
-    let action = csa_scheduler::decide_failover(
-        tool_name_str,
-        "default",
-        resolved_tier_name,
-        task_needs_edit,
-        session_state.as_ref(),
-        tried_tools,
-        tried_specs,
-        cfg,
-        &rate_limit.matched_pattern,
-    );
-
-    match action {
-        csa_scheduler::FailoverAction::RetryInSession {
-            new_tool,
-            new_model_spec,
-            session_id: _,
-        }
-        | csa_scheduler::FailoverAction::RetrySiblingSession {
-            new_tool,
-            new_model_spec,
-        } => {
-            warn!(
-                from_tool = %tool_name_str,
-                from_spec = %current_model_spec.unwrap_or("none"),
-                to_tool = %new_tool,
-                to_spec = %new_model_spec,
-                quota_exhausted = rate_limit.quota_exhausted,
-                reason = %rate_limit.reason,
-                "[csa-failover] intra-tier failover"
-            );
-            fallback_chain.push(FallbackAttempt {
-                tool: tool_name_str.to_string(),
-                model_spec: current_model_spec.map(String::from),
-                skip_reason: rate_limit.matched_pattern.clone(),
-                quota_exhausted: rate_limit.quota_exhausted,
-                timestamp: chrono::Utc::now(),
-            });
-            let tool = crate::run_helpers::parse_tool_name(&new_tool)?;
-            Ok(RateLimitAction::Retry {
-                new_tool: tool,
-                new_model_spec: Some(new_model_spec),
-            })
-        }
-        csa_scheduler::FailoverAction::ReportError { reason, .. } => {
-            warn!(reason = %reason, "Failover not possible, returning original result");
-            Ok(RateLimitAction::ExhaustedFailovers)
-        }
-    }
-}
-
-/// Check an anyhow error message for rate-limit signals and attempt failover.
-///
-/// This handles the case where the execution returned `Err(e)` (e.g. ACP
-/// `PromptFailed` with `usage_limit_exceeded`) instead of a non-zero
-/// `ExecutionResult`. The error text is tested against the same rate-limit
-/// patterns used for normal exit-code-based detection.
-/// Appends a `FallbackAttempt` to `fallback_chain` when a retry is triggered.
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn evaluate_error_rate_limit_failover(
-    tool_name_str: &str,
-    error_message: &str,
-    attempts: usize,
-    max_failover_attempts: usize,
-    tried_tools: &mut Vec<String>,
-    tried_specs: &mut Vec<String>,
-    tier_auto_select: bool,
-    failover_on_crash_enabled: bool,
-    resolved_tier_name: Option<&str>,
-    executed_session_id: Option<&str>,
-    effective_session_arg: Option<&str>,
-    ephemeral: bool,
-    prompt_text: &str,
-    project_root: &Path,
-    config: Option<&ProjectConfig>,
-    task_needs_edit: Option<bool>,
-    current_model_spec: Option<&str>,
-    fallback_chain: &mut FallbackChain,
-    attempt_elapsed: Option<Duration>,
-) -> Result<RateLimitAction> {
-    let failover_signal = match detect_transport_error_failover_signal(
-        tool_name_str,
-        error_message,
-        current_model_spec,
-    ) {
-        Some(signal) => signal,
-        None => return Ok(RateLimitAction::NoRateLimit),
-    };
-
-    if failover_signal.quota_exhausted {
-        warn!(
-            tool = %tool_name_str,
-            pattern = %failover_signal.matched_pattern,
-            "Permanent tool quota exhaustion detected in transport error; not attempting runtime failover"
-        );
-        return Ok(RateLimitAction::ExhaustedFailovers);
-    }
-
-    if !allows_init_failure_failover(
-        tool_name_str,
-        &failover_signal.reason,
-        failover_signal.requires_init_failure_window,
-        attempt_elapsed,
-    ) {
-        return Ok(RateLimitAction::NoRateLimit);
-    }
-
-    match failover_signal.kind {
-        TransportErrorFailoverKind::RateLimit => {
-            if !tier_auto_select {
-                return Ok(RateLimitAction::NoRateLimit);
-            }
-            info!(
-                tool = %tool_name_str,
-                pattern = %failover_signal.matched_pattern,
-                attempt = attempts,
-                max = max_failover_attempts,
-                "Rate limit detected in transport error, attempting failover"
-            );
-        }
-        TransportErrorFailoverKind::AcpCrashRetryExhausted => {
-            if !failover_on_crash_enabled {
-                return Ok(RateLimitAction::NoRateLimit);
-            }
-            warn!(
-                tool = %tool_name_str,
-                pattern = %failover_signal.matched_pattern,
-                attempt = attempts,
-                max = max_failover_attempts,
-                "[csa-failover] ACP crash retry exhaustion detected in transport error; attempting tier failover"
-            );
-        }
-        TransportErrorFailoverKind::GeminiRetryChainExhausted => {
-            if !failover_on_crash_enabled {
-                return Ok(RateLimitAction::NoRateLimit);
-            }
-            warn!(
-                tool = %tool_name_str,
-                pattern = %failover_signal.matched_pattern,
-                attempt = attempts,
-                max = max_failover_attempts,
-                "[csa-failover] Gemini retry chain exhaustion detected in transport error; attempting tier failover"
-            );
-        }
-    }
-
-    if attempts >= max_failover_attempts {
-        warn!(
-            "Max failover attempts ({}) reached for error-path rate limit",
-            max_failover_attempts
-        );
-        return Ok(RateLimitAction::ExhaustedFailovers);
-    }
-
-    tried_tools.push(tool_name_str.to_string());
-    if let Some(spec) = current_model_spec {
-        tried_specs.push(spec.to_string());
-    }
-
-    let failover_session_ref = executed_session_id.or(effective_session_arg);
-    let session_state = if !ephemeral {
-        failover_session_ref.and_then(|sid| {
-            let sessions_dir = csa_session::get_session_root(project_root)
-                .ok()?
-                .join("sessions");
-            let resolved_id = csa_session::resolve_session_prefix(&sessions_dir, sid).ok()?;
-            csa_session::load_session(project_root, &resolved_id).ok()
-        })
-    } else {
-        None
-    };
-
-    let task_needs_edit = task_needs_edit
-        .or_else(|| crate::run_helpers::infer_task_edit_requirement(prompt_text))
-        .or_else(|| config.map(|cfg| cfg.can_tool_edit_existing(tool_name_str)));
-
-    let Some(cfg) = config else {
-        return Ok(RateLimitAction::ExhaustedFailovers);
-    };
-
-    let action = csa_scheduler::decide_failover(
-        tool_name_str,
-        "default",
-        resolved_tier_name,
-        task_needs_edit,
-        session_state.as_ref(),
-        tried_tools,
-        tried_specs,
-        cfg,
-        &failover_signal.matched_pattern,
-    );
-
-    match action {
-        csa_scheduler::FailoverAction::RetryInSession {
-            new_tool,
-            new_model_spec,
-            session_id: _,
-        }
-        | csa_scheduler::FailoverAction::RetrySiblingSession {
-            new_tool,
-            new_model_spec,
-        } => {
-            warn!(
-                from_tool = %tool_name_str,
-                from_spec = %current_model_spec.unwrap_or("none"),
-                to_tool = %new_tool,
-                to_spec = %new_model_spec,
-                quota_exhausted = failover_signal.quota_exhausted,
-                reason = %failover_signal.reason,
-                "[csa-failover] intra-tier failover (transport error)"
-            );
-            fallback_chain.push(FallbackAttempt {
-                tool: tool_name_str.to_string(),
-                model_spec: current_model_spec.map(String::from),
-                skip_reason: failover_signal.matched_pattern.clone(),
-                quota_exhausted: failover_signal.quota_exhausted,
-                timestamp: chrono::Utc::now(),
-            });
-            let tool = crate::run_helpers::parse_tool_name(&new_tool)?;
-            Ok(RateLimitAction::Retry {
-                new_tool: tool,
-                new_model_spec: Some(new_model_spec),
-            })
-        }
-        csa_scheduler::FailoverAction::ReportError { reason, .. } => {
-            warn!(reason = %reason, "Error-path failover not possible");
-            Ok(RateLimitAction::ExhaustedFailovers)
         }
     }
 }
